// frontend/src/api/generated/models/publicLinkCreateScope.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * 途排智策 API
 * 途排智策前后端分离 MVP API
 * OpenAPI spec version: 0.1.0
 */

export type PublicLinkCreateScope = typeof PublicLinkCreateScope[keyof typeof PublicLinkCreateScope];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PublicLinkCreateScope = {
  class: 'class',
  teacher: 'teacher',
  school: 'school',
} as const;