// frontend/src/api/generated/models/publicLinkSecretResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * 途排智策 API
 * 途排智策前后端分离 MVP API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicLinkSecretResponseCampusId } from './publicLinkSecretResponseCampusId';
import type { PublicLinkSecretResponseResourceBusinessId } from './publicLinkSecretResponseResourceBusinessId';
import type { PublicLinkSecretResponseExpiresAt } from './publicLinkSecretResponseExpiresAt';
import type { PublicLinkSecretResponseLastSeenAt } from './publicLinkSecretResponseLastSeenAt';
import type { PublicLinkSecretResponseCreatedBy } from './publicLinkSecretResponseCreatedBy';

export interface PublicLinkSecretResponse {
  id: string;
  schedule_set_id: string;
  scope: string;
  campus_id: PublicLinkSecretResponseCampusId;
  resource_business_id: PublicLinkSecretResponseResourceBusinessId;
  display_name: string;
  show_teacher_names: boolean;
  token_hint: string;
  status: string;
  expires_at: PublicLinkSecretResponseExpiresAt;
  last_seen_at: PublicLinkSecretResponseLastSeenAt;
  access_count: number;
  note: string;
  created_at: string;
  created_by: PublicLinkSecretResponseCreatedBy;
  token: string;
  public_url: string;
}